service/requirements.txt
Django>=3.0,<4.0
psycopg2==2.9.1
django-bootstrap5
django-random-user-hash==0.0.5
gunicorn>=20.0
Pillow==9.1.1
django-crontab
python-memcached
PyCryptoDome
